+++ .env.example
@@ -7,7 +7,7 @@ PORT=8081
 BROWSER=none
 
 
-# App dotenv. These variables are used DURING react build.
+# Page dotenv. These variables are used DURING react build.
 # For this reason, the values below will only be used with
 # `pnpm dev`. For values used along with `pnpm build`, check
 # the Dockerfile (docker build arguments)
