--- src/app/page.js
@@ -1,6 +1,0 @@
-"use client";
-import React from 'react';
-import ReactDOM from 'react-dom';
-import App from '../app';
-
-ReactDOM.render(<App />, document.getElementById('root'));
