+++ src/components/modals/modal.tsx
@@ -9,8 +9,6 @@ interface ModalProps {
   children: React.ReactNode;
 }
 
-ModalComponent.setAppElement('#root');
-
 const Modal = ({ isOpen, className, children }: ModalProps) => {
   return (
     <ModalComponent isOpen={isOpen} className={`modal ${className ? className : ''}`} overlayClassName="overlay">
